requirements.txt
Flask
Gunicorn
flask_cors
python-dotenv
pystrapi
typing_extensions